refactor: cleanup properties module
