--- properties/src/main/java/love/broccolai/corn/properties/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Corn property module
- */
-package love.broccolai.corn.properties;
