+++ properties/src/main/java/love/broccolai/corn/properties/property/NestedProperty.java
@@ -1,5 +1,6 @@
-package love.broccolai.corn.properties;
+package love.broccolai.corn.properties.property;
 
+import love.broccolai.corn.properties.PropertyHolder;
 import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.Nullable;
