+++ properties/src/main/java/love/broccolai/corn/properties/property/ObjectProperty.java
@@ -1,0 +1,9 @@
+package love.broccolai.corn.properties.property;
+
+import org.jspecify.annotations.NullMarked;
+import org.jspecify.annotations.Nullable;
+
+@NullMarked
+record ObjectProperty(String name, @Nullable Object object) implements Property {
+
+}
