+++ properties/src/main/java/love/broccolai/corn/properties/snapshot/PropertySnapshotImpl.java
@@ -1,5 +1,7 @@
-package love.broccolai.corn.properties;
+package love.broccolai.corn.properties.snapshot;
 
+import love.broccolai.corn.properties.property.FlattenableProperty;
+import love.broccolai.corn.properties.property.Property;
 import org.jspecify.annotations.NullMarked;
 
 import java.util.Collection;
